requirements.txt
pandas>=2.0.0
requests>=2.28.0
python-dotenv>=1.0.0
pyarrow>=14.0.0
mysql-connector-python>=8.0.0
tzdata>=2023.3
